move dispatches to cli/js/ops/
This commit moves cli/js/dispatch_json.ts and cli/js/dispatch_minimal.ts
to cli/js/ops/ in an effort to factor out code dedicated to JS side of
ops to replicate directory structure in cli/ops/.
