+++ cli/js/format_error.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 import { DiagnosticItem } from "./diagnostics.ts";
-import { sendSync } from "./dispatch_json.ts";
+import { sendSync } from "./ops/dispatch_json.ts";
 
 /**
  * Format an array of diagnostic items and return them as a single string.
