+++ cli/js/get_random_values.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { sendSync } from "./dispatch_json.ts";
+import { sendSync } from "./ops/dispatch_json.ts";
 import { assert } from "./util.ts";
 
 /** Synchronously collects cryptographically secure random values. The
