+++ cli/js/ops/dispatch_minimal.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import * as util from "./util.ts";
-import { core } from "./core.ts";
-import { TextDecoder } from "./web/text_encoding.ts";
-import { ErrorKind, errors, getErrorClass } from "./errors.ts";
+import * as util from "../util.ts";
+import { core } from "../core.ts";
+import { TextDecoder } from "../web/text_encoding.ts";
+import { ErrorKind, errors, getErrorClass } from "../errors.ts";
 
 const promiseTableMin = new Map<number, util.Resolvable<RecordMinimal>>();
 // Note it's important that promiseId starts at 1 instead of 0, because sync
