+++ cli/js/read_dir.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { sendSync, sendAsync } from "./dispatch_json.ts";
+import { sendSync, sendAsync } from "./ops/dispatch_json.ts";
 import { FileInfo, FileInfoImpl } from "./file_info.ts";
 import { StatResponse } from "./stat.ts";
