+++ cli/js/runtime.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 import { core } from "./core.ts";
 import * as dispatch from "./dispatch.ts";
-import { sendSync } from "./dispatch_json.ts";
+import { sendSync } from "./ops/dispatch_json.ts";
 import { assert } from "./util.ts";
 import * as util from "./util.ts";
 import { OperatingSystem, Arch } from "./build.ts";
