+++ cli/js/web/fetch.ts
@@ -15,7 +15,7 @@ import { Buffer } from "../buffer.ts";
 import { FormData } from "./form_data.ts";
 import { URL } from "./url.ts";
 import { URLSearchParams } from "./url_search_params.ts";
-import { sendAsync } from "../dispatch_json.ts";
+import { sendAsync } from "../ops/dispatch_json.ts";
 
 function getHeaderValueParams(value: string): Map<string, string> {
   const params = new Map();
