+++ cli/tests/error_012_bad_dynamic_import_specifier.ts.out
@@ -1,7 +1,7 @@
 [WILDCARD]error: Uncaught URIError: relative import path "bad-module.ts" not prefixed with / or ./ or ../ Imported from "[WILDCARD]/error_012_bad_dynamic_import_specifier.ts"
 [WILDCARD]dispatch_json.ts:[WILDCARD]
-    at unwrapResponse ($deno$/dispatch_json.ts:[WILDCARD])
-    at sendSync ($deno$/dispatch_json.ts:[WILDCARD])
+    at unwrapResponse ($deno$/ops/dispatch_json.ts:[WILDCARD])
+    at sendSync ($deno$/ops/dispatch_json.ts:[WILDCARD])
     at resolveModules ($deno$/compiler_imports.ts:[WILDCARD])
     at processImports ($deno$/compiler_imports.ts:[WILDCARD])
     at processImports ($deno$/compiler_imports.ts:[WILDCARD])
