+++ cli/tests/error_type_definitions.ts.out
@@ -1,7 +1,7 @@
 [WILDCARD]error: Uncaught URIError: relative import path "baz" not prefixed with / or ./ or ../ Imported from "[WILDCARD]/type_definitions/bar.d.ts"
 [WILDCARD]dispatch_json.ts:[WILDCARD]
-    at unwrapResponse ($deno$/dispatch_json.ts:[WILDCARD])
-    at sendSync ($deno$/dispatch_json.ts:[WILDCARD])
+    at unwrapResponse ($deno$/ops/dispatch_json.ts:[WILDCARD])
+    at sendSync ($deno$/ops/dispatch_json.ts:[WILDCARD])
     at resolveModules ($deno$/compiler_imports.ts:[WILDCARD])
     at processImports ($deno$/compiler_imports.ts:[WILDCARD])
     at processImports ($deno$/compiler_imports.ts:[WILDCARD])
